+++ hydrolib/tools/extforce_convert/main_converter.py
@@ -7,6 +7,7 @@ from typing import Any, Dict, Optional, Tuple, Union
 from tqdm import tqdm
 
 from hydrolib.core.base.file_manager import PathOrStr
+from hydrolib.core.base.utils import PathStyle
 from hydrolib.core.dflowfm.ext.models import (
     Boundary,
     ExtModel,
@@ -24,6 +25,7 @@ from hydrolib.core.dflowfm.structure.models import Structure, StructureModel
 from hydrolib.tools.extforce_convert.converters import ConverterFactory
 from hydrolib.tools.extforce_convert.mdu_parser import MDUParser
 from hydrolib.tools.extforce_convert.utils import (
+    CONVERTER_DATA,
     backup_file,
     construct_filemodel_new_or_existing,
 )
@@ -40,6 +42,8 @@ class ExternalForcingConverter:
         structure_file: Optional[PathOrStr] = None,
         mdu_parser: MDUParser = None,
         verbose: bool = False,
+        path_style: PathStyle = None,
+        debug: Optional[bool] = False,
     ):
         r"""Initialize the converter.
 
@@ -60,6 +64,15 @@ class ExternalForcingConverter:
                 a Parser for the MDU file.
             verbose (bool, optional, Defaults to False):
                 Enable verbose output.
+            path_style (Optional[PathStyle], optional):
+                Specifies how absolute paths in model files are interpreted and converted (unix or windows style).
+                Use this to correctly handle Unix paths on Windows or Windows paths on Unix.
+                If None, the system's default style is used.
+                - Example (Windows): `P:/path/to/file`
+                - Example (Unix): `/p/path/to/file`
+            debug (bool, Optional):
+                Enable debug mode. In debug mode unsupported quantities will be skipped and not raise an error.
+                Defaults to False.
 
         Raises:
             FileNotFoundError: If the old external forcing file does not exist.
@@ -91,7 +104,7 @@ class ExternalForcingConverter:
             ```
         """
         if isinstance(extold_model, Path) or isinstance(extold_model, str):
-            extold_model = self._read_old_file(extold_model)
+            extold_model = self._read_old_file(extold_model, path_style=path_style)
             rdir = extold_model.filepath.parent
         else:
             if not isinstance(extold_model, ExtOldModel):
@@ -103,6 +116,7 @@ class ExternalForcingConverter:
         self._extold_model = extold_model
         self._verbose = verbose
         self._root_dir = rdir
+        self._path_style = path_style
 
         # create the new models if not provided by the user in the same directory as the old external file
         path = rdir / "new-external-forcing.ext" if ext_file is None else ext_file
@@ -131,6 +145,19 @@ class ExternalForcingConverter:
             self._mdu_parser = mdu_parser
 
         self._legacy_files = []
+        self.debug = debug
+        self.un_supported_quantities = self.check_unsupported_quantities()
+
+    def check_unsupported_quantities(self):
+        """Check for unsupported quantities in the old external forcing model.
+
+        Returns:
+            list[str]: List of unsupported quantities found in the old external forcing model.
+        """
+        # check if the file contains unsupported quantities
+        return CONVERTER_DATA.check_unsupported_quantities(
+            self.extold_model, raise_error=not self.debug
+        )
 
     @property
     def legacy_files(self) -> list[Path]:
@@ -165,6 +192,11 @@ class ExternalForcingConverter:
     def root_dir(self) -> Path:
         """Root directory of the external forcing file."""
         return self._root_dir
+
+    @property
+    def path_style(self) -> PathStyle:
+        """Path style for the absolute paths in the models."""
+        return self._path_style
 
     @property
     def extold_model(self) -> ExtOldModel:
@@ -218,12 +250,17 @@ class ExternalForcingConverter:
         )
 
     @staticmethod
-    def _read_old_file(extoldfile: PathOrStr) -> ExtOldModel:
+    def _read_old_file(
+        ext_old_file: PathOrStr, path_style: Optional[PathStyle] = None
+    ) -> ExtOldModel:
         """Read a legacy D-Flow FM external forcings file (.ext) into an ExtOldModel object.
 
         Args:
-            extoldfile (PathOrStr):
+            ext_old_file (PathOrStr):
                 path to the external forcings file (.ext)
+            path_style (Optional[PathStyle]):
+                Path style for the file paths in the models. If None, the system style is used.
+                Converts absolute paths based on the provided style to the system style.
 
             Returns:
                 ExtOldModel: object with all forcing blocks.
@@ -231,15 +268,15 @@ class ExternalForcingConverter:
             Raises:
                 FileNotFoundError: If the old external forcing file does not exist.
         """
-        if not isinstance(extoldfile, Path):
-            extoldfile = Path(extoldfile)
+        if not isinstance(ext_old_file, Path):
+            ext_old_file = Path(ext_old_file)
 
-        if not extoldfile.exists():
-            raise FileNotFoundError(f"File not found: {extoldfile}")
+        if not ext_old_file.exists():
+            raise FileNotFoundError(f"File not found: {ext_old_file}")
 
-        extold_model = ExtOldModel(extoldfile)
+        ext_old_model = ExtOldModel(ext_old_file, path_style=path_style)
 
-        return extold_model
+        return ext_old_model
 
     def _type_field_map(self) -> dict[type, tuple[Any, str]]:
         return {
@@ -274,6 +311,14 @@ class ExternalForcingConverter:
             total=num_quantities, desc="Converting forcings", unit="forcing"
         ) as progress_bar:
             for forcing in self.extold_model.forcing:
+                if forcing.quantity in self.un_supported_quantities:
+                    print(
+                        f"The quantity {forcing.quantity} is not supported, and the debug mode is {self.debug}. "
+                        "So the forcing will not be converted (stay in the old external forcing file)."
+                    )
+                    progress_bar.update(1)
+                    continue
+
                 # Update the description of tqdm to include the current forcing's filepath
                 progress_bar.set_description(
                     f"Processing: {forcing.quantity} - {forcing.filename.filepath}"
@@ -295,6 +340,14 @@ class ExternalForcingConverter:
 
         if self.mdu_parser is not None:
             self._update_mdu_file()
+
+        if self.un_supported_quantities:
+            forcing = [
+                forcing
+                for forcing in self.extold_model.forcing
+                if forcing.quantity in self.un_supported_quantities
+            ]
+            self.extold_model.forcing = forcing
 
         return self.ext_model, self.inifield_model, self.structure_model
 
@@ -355,6 +408,10 @@ class ExternalForcingConverter:
         if len(self.structure_model.structure) > 0:
             self._save_structure_model(backup, recursive)
 
+        if self.un_supported_quantities:
+            backup_file(self.extold_model.filepath)
+            self.extold_model.save(recurse=recursive, exclude_unset=True)
+
         num_quantities_ext = (
             len(self.ext_model.meteo)
             + len(self.ext_model.sourcesink)
@@ -364,7 +421,9 @@ class ExternalForcingConverter:
         if num_quantities_ext:
             if backup and self.ext_model.filepath.exists():
                 backup_file(self.ext_model.filepath)
-            self.ext_model.save(recurse=recursive, exclude_unset=True)
+            self.ext_model.save(
+                recurse=recursive, exclude_unset=True, path_style=self.path_style
+            )
 
         if self.mdu_parser is not None:
             self.mdu_parser.clean()
@@ -378,12 +437,16 @@ class ExternalForcingConverter:
         """
         if backup and self.inifield_model.filepath.exists():
             backup_file(self.inifield_model.filepath)
-        self.inifield_model.save(recurse=recursive, exclude_unset=False)
+        self.inifield_model.save(
+            recurse=recursive, exclude_unset=False, path_style=self.path_style
+        )
 
     def _save_structure_model(self, backup: bool, recursive: bool):
         if backup and self.structure_model.filepath.exists():
             backup_file(self.structure_model.filepath)
-        self.structure_model.save(recurse=recursive, exclude_unset=True)
+        self.structure_model.save(
+            recurse=recursive, exclude_unset=True, path_style=self.path_style
+        )
 
     def clean(self):
         """Clean the directory from the old external forcing file and the time file."""
@@ -392,7 +455,8 @@ class ExternalForcingConverter:
                 print(f"Removing legacy file:{file}")
                 file.unlink()
 
-        self.extold_model.filepath.unlink()
+        if not self.un_supported_quantities:
+            self.extold_model.filepath.unlink()
 
     @classmethod
     def from_mdu(
@@ -401,6 +465,8 @@ class ExternalForcingConverter:
         ext_file_user: Optional[PathOrStr] = None,
         inifield_file_user: Optional[PathOrStr] = None,
         structure_file_user: Optional[PathOrStr] = None,
+        path_style: Optional[PathStyle] = None,
+        debug: bool = False,
     ) -> "ExternalForcingConverter":
         """Create the converter from the MDU file.
 
@@ -418,6 +484,12 @@ class ExternalForcingConverter:
             structure_file_user (PathOrStr, optional): Path to the output structures.ini
                 file. Defaults to the given StructureFile in the MDU file, if
                 present, or structures.ini otherwise.
+            path_style (Optional[PathStyle], optional):
+                Path style for the file paths in the models. If None, the system style is used.
+                Converts absolute paths based on the provided style to the system style.
+            debug (bool, Optional):
+                Enable debug mode. In debug mode unsupported quantities will be skipped and not raise an error.
+                Defaults to False.
 
         Returns:
             ExternalForcingConverter: The converter object.
@@ -443,6 +515,8 @@ class ExternalForcingConverter:
             inifield_file_user,
             structure_file_user,
             mdu_parser,
+            path_style=path_style,
+            debug=debug,
         )
 
     def _update_mdu_file(self):
@@ -463,8 +537,12 @@ class ExternalForcingConverter:
             + len(self.ext_model.sourcesink)
         )
 
+        remove_old_ext_file = False if self.un_supported_quantities else True
+
         self.mdu_parser.update_extforce_file_new(
-            self.ext_model.filepath.name, num_quantities=num_ext_model_quantities
+            self.ext_model.filepath.name,
+            num_quantities=num_ext_model_quantities,
+            remove_old_ext_file=remove_old_ext_file,
         )
 
         num_quantities_inifield = len(self.inifield_model.parameter) + len(
@@ -496,6 +574,7 @@ def recursive_converter(
     backup: bool = True,
     suppress_errors: bool = False,
     remove_legacy: bool = False,
+    debug: bool = False,
 ):
     """Migrate all external forcings files in a directory tree to the new format.
 
@@ -504,6 +583,9 @@ def recursive_converter(
         backup (bool, optional): Create a backup of each file that will be overwritten.
         suppress_errors (bool, optional): Suppress errors during conversion.
         remove_legacy (bool, optional): Remove legacy/old files (e.g. .tim) after conversion. Defaults to False.
+        debug (bool, Optional):
+                Enable debug mode. In debug mode unsupported quantities will be skipped and not raise an error.
+                Defaults to False.
     """
     mdu_files = [
         path for path in Path(root_dir).rglob("*.mdu") if "_ext" not in path.name
@@ -516,7 +598,7 @@ def recursive_converter(
 
         for path in tqdm(mdu_files, desc="Converting files"):
             try:
-                converter = ExternalForcingConverter.from_mdu(path)
+                converter = ExternalForcingConverter.from_mdu(path, debug=debug)
                 _, _, _ = converter.update()
                 converter.save(backup=backup)
                 if remove_legacy: